ports/playstation2: use Makefile.iopglobal from PS2SDK
This now results in a valid IRX file that can be loaded and executed
through the loadfile mechanism.

Before this change, SifExecModuleBuffer() with the underlying call of
sceSifCallRpc() would fail, returning 201 for "Invalid IRX module".

Analysis of the previously generated .irx file showed that it did not
have the properties that make an ELF file a valid IRX file.